requirements.txt
# Python 3.11

deep-translator>=1.11.4
email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
pdf2image>=1.17.0
pillow>=11.2.1
psycopg2-binary>=2.9.10
pytesseract>=0.3.13
requests>=2.32.3
werkzeug>=3.1.3